[update]: separate ros2usb into a single class file
