+++ src/ros2usb.cpp
@@ -131,21 +131,3 @@ void ROS2USB::sendToNode() {
   memcpy(msg.data.data(), buf.data() + 3, len - 5);
   publisher_->publish(msg);
 }
-
-int main(int argc, char **argv) {
-  rclcpp::init(argc, argv);
-  auto ros2usb_node = std::make_shared<ROS2USB>();
-
-  try {
-    RCLCPP_INFO(ros2usb_node->get_logger(), "Initializing node...");
-    ros2usb_node->config();
-    RCLCPP_INFO(ros2usb_node->get_logger(), "Node initialized");
-    rclcpp::spin(ros2usb_node);
-  } catch (const char *s) {
-    RCLCPP_FATAL_STREAM(ros2usb_node->get_logger(), "" << s);
-  } catch (...) {
-    RCLCPP_FATAL_STREAM(ros2usb_node->get_logger(), "Unexpected error");
-  }
-
-  return 0;
-}
